requirements.txt
asgiref==3.5.2; python_version >= "3.8"
django-environ==0.9.0; python_version >= "3.4" and python_version < "4"
django-extensions==3.2.0; python_version >= "3.6"
django-phonenumber-field==6.3.0; python_version >= "3.7"
django==4.0.6; python_version >= "3.8"
djangorestframework==3.13.1; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
phonenumbers==8.12.51
pillow==9.2.0; python_version >= "3.7"
psycopg2-binary==2.9.3; python_version >= "3.6"
pytz==2022.1; python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2022.1; sys_platform == "win32" and python_version >= "3.8"
